Complete removing evictionpolicy separation, merging all eviction logic into listgauge
